Move GenerateOrderGuidAsync to IOrderProcessingService
