+++ src/Presentation/Nop.Web/Controllers/CheckoutController.cs
@@ -552,9 +552,9 @@ public partial class CheckoutController : BasePublicController
                     redirect = Url.RouteUrl("CheckoutShippingAddress")
                 });
 
-            var shippingAddressModel  = new CheckoutShippingAddressModel();
+            var shippingAddressModel = new CheckoutShippingAddressModel();
             await _checkoutModelFactory.PrepareShippingAddressModelAsync(shippingAddressModel, cart, address.CountryId);
-            
+
             return Json(new
             {
                 selected_id = model.ShippingNewAddress.Id,
@@ -766,7 +766,7 @@ public partial class CheckoutController : BasePublicController
         //model
         var model = new CheckoutShippingAddressModel();
         await _checkoutModelFactory.PrepareShippingAddressModelAsync(model, cart, prePopulateNewAddressWithCustomerFields: true);
-        
+
         return View(model);
     }
 
@@ -881,7 +881,7 @@ public partial class CheckoutController : BasePublicController
         await _checkoutModelFactory.PrepareShippingAddressModelAsync(model, cart,
             selectedCountryId: newAddress.CountryId,
             overrideAttributesXml: customAttributes);
-        
+
         return View(model);
     }
 
@@ -1220,7 +1220,7 @@ public partial class CheckoutController : BasePublicController
             //get payment info
             var paymentInfo = await paymentMethod.GetPaymentInfoAsync(form);
             //set previous order GUID (if exists)
-            await _paymentService.GenerateOrderGuidAsync(paymentInfo);
+            await _orderProcessingService.GenerateOrderGuidAsync(paymentInfo);
 
             //session save
             await HttpContext.Session.SetAsync("OrderPaymentInfo", paymentInfo);
@@ -1307,7 +1307,7 @@ public partial class CheckoutController : BasePublicController
 
                 processPaymentRequest = new ProcessPaymentRequest();
             }
-            await _paymentService.GenerateOrderGuidAsync(processPaymentRequest);
+            await _orderProcessingService.GenerateOrderGuidAsync(processPaymentRequest);
             processPaymentRequest.StoreId = store.Id;
             processPaymentRequest.CustomerId = customer.Id;
             processPaymentRequest.PaymentMethodSystemName = await _genericAttributeService.GetAttributeAsync<string>(customer,
@@ -1565,7 +1565,7 @@ public partial class CheckoutController : BasePublicController
                         selectedCountryId: newAddress.CountryId,
                         overrideAttributesXml: customAttributes);
                     billingAddressModel.NewAddressPreselected = true;
-                    
+
                     return Json(new
                     {
                         update_section = new UpdateSectionJsonModel
@@ -1733,7 +1733,7 @@ public partial class CheckoutController : BasePublicController
                         selectedCountryId: newAddress.CountryId,
                         overrideAttributesXml: customAttributes);
                     shippingAddressModel.NewAddressPreselected = true;
-                    
+
                     return Json(new
                     {
                         update_section = new UpdateSectionJsonModel
@@ -1972,7 +1972,7 @@ public partial class CheckoutController : BasePublicController
                 //get payment info
                 var paymentInfo = await paymentMethod.GetPaymentInfoAsync(form);
                 //set previous order GUID (if exists)
-                await _paymentService.GenerateOrderGuidAsync(paymentInfo);
+                await _orderProcessingService.GenerateOrderGuidAsync(paymentInfo);
 
                 //session save
                 await HttpContext.Session.SetAsync("OrderPaymentInfo", paymentInfo);
@@ -2058,7 +2058,7 @@ public partial class CheckoutController : BasePublicController
 
                     processPaymentRequest = new ProcessPaymentRequest();
                 }
-                await _paymentService.GenerateOrderGuidAsync(processPaymentRequest);
+                await _orderProcessingService.GenerateOrderGuidAsync(processPaymentRequest);
                 processPaymentRequest.StoreId = store.Id;
                 processPaymentRequest.CustomerId = customer.Id;
                 processPaymentRequest.PaymentMethodSystemName = await _genericAttributeService.GetAttributeAsync<string>(customer,
